+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- New npm dependency `q` to create deferred 
+- New npm dependency `slash` to rewrite a path with slash instead of backslash
+- New input `validate` to preview and confirm the creation of a service
+- New input `edit` called when validate is false. Used to list inputs that could be edited. On select, execute the input with the default value as the value from the config. On submit, restart to with `validate` input. Fucking hell of a loop of deferred !
+
+### Changed
+- Removed all replace to by slash module to rewrite a path with slash instead of backslash 
 
 ## [0.12.0]
 ### Added